+++ CHANGELOG.md
@@ -55,7 +55,8 @@
 * Tests for route `POST /cards`.
 * Tests for route `PATCH /cards/:id`.
 * Tests for route `DELETE /cards/:id`.
-* Tests for route `GET /games`. 
+* Tests for route `GET /games`.
+* Tests for route `GET /games/:id`.
 * Tests for route `POST /games`.
 
 ### 📦 Packages